Add files via upload
In this captivating Power BI project, data sourced from SQL brings to life the dynamic world of IPL cricket. Delve into the performances of the top 10 batsmen and bowlers from 2008 to 2023 with stunning visuals and insightful analysis. Witness their runs, wickets, averages, and strike rates unfold through interactive charts, slicers, and scrollers. Navigate seamlessly through the years, comparing player statistics, trends, and achievements with ease. Immerse yourself in the glory of IPL cricket history, as Power BI weaves a narrative of excellence, passion, and skill, showcasing the brilliance of the game's finest talents in a visually captivating and engaging manner.